refactor: concurrent search, file cohesion, error convention
R1: Replace sequential fetchSkillMeta loop with errgroup (SetLimit 4)
R2: Extract parseSkillFrontmatter to frontmatter.go; move ghRepoURL,
    findSkillDir, findRepoRoot to manager_impl.go; handlers become pure
    thin controllers with no helper functions
R3: Change error returns from fmt.Sprintf('Error: %v', err), err to
    '', fmt.Errorf('context: %w', err); handlers format Text from err
