+++ internal/tools/integrations/skillssh/install.go
@@ -6,14 +6,9 @@ package skillssh
 import (
 	"context"
 	"fmt"
-	"regexp"
 
 	"github.com/gosharplite/tell-me-go/internal/domain/tools"
 )
-
-// ghRepoURL matches GitHub repository URLs: https://github.com/<owner>/<repo>
-// Optional trailing slash and .git suffix are stripped during normalization.
-var ghRepoURL = regexp.MustCompile(`^https://github\.com/([^/]+)/([^/]+?)(?:\.git)?/?$`)
 
 // makeInstallSkill returns a handler for the install_skill tool.
 func makeInstallSkill(mgr SkillManager) tools.ToolFunc {
@@ -27,7 +22,7 @@ func makeInstallSkill(mgr SkillManager) tools.ToolFunc {
 
 		output, err := mgr.InstallSkill(ctx, params.RepoURL)
 		if err != nil {
-			return tools.ToolResult{Error: err, Text: output}, nil
+			return tools.ToolResult{Error: err, Text: fmt.Sprintf("Error: %v", err)}, nil
 		}
 		return tools.ToolResult{Text: output}, nil
 	}
